requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.2
    # via
    #   fsspec
    #   langchain-community
aiosignal==1.3.2
    # via aiohttp
altair==4.0.0
    # via
    #   -r requirements/requirements.in
    #   streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
appdirs==1.4.4
    # via ragas
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via streamlit
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via streamlit
dataclasses-json==0.6.7
    # via langchain-community
datasets==3.6.0
    # via ragas
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
diskcache==5.6.3
    # via ragas
distro==1.9.0
    # via openai
entrypoints==0.4
    # via altair
faiss-cpu==1.11.0
    # via -r requirements/requirements.in
filelock==3.18.0
    # via
    #   datasets
    #   huggingface-hub
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2025.3.0
    # via
    #   datasets
    #   huggingface-hub
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
h11==0.16.0
    # via httpcore
hf-xet==1.1.2
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langsmith
    #   openai
httpx-sse==0.4.0
    # via langchain-community
huggingface-hub==0.32.1
    # via datasets
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jiter==0.10.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.24.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
langchain==0.3.25
    # via
    #   langchain-community
    #   ragas
langchain-community==0.3.24
    # via
    #   -r requirements/requirements.in
    #   ragas
langchain-core==0.3.61
    # via
    #   -r requirements/requirements.in
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
    #   ragas
langchain-openai==0.3.0
    # via
    #   -r requirements/requirements.in
    #   ragas
langchain-text-splitters==0.3.8
    # via langchain
langsmith==0.3.42
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
load-dotenv==0.1.0
    # via -r requirements/requirements.in
longchain==0.4.2
    # via -r requirements/requirements.in
markupsafe==3.0.2
    # via jinja2
marshmallow==3.26.1
    # via dataclasses-json
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
mypy-extensions==1.1.0
    # via typing-inspect
nest-asyncio==1.6.0
    # via ragas
numpy==2.2.6
    # via
    #   altair
    #   datasets
    #   faiss-cpu
    #   langchain-community
    #   pandas
    #   pydeck
    #   ragas
    #   streamlit
openai==1.82.0
    # via
    #   langchain-openai
    #   ragas
orjson==3.10.18
    # via langsmith
packaging==24.2
    # via
    #   datasets
    #   faiss-cpu
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   marshmallow
    #   streamlit
pandas==2.2.3
    # via
    #   altair
    #   datasets
    #   streamlit
pillow==11.2.1
    # via streamlit
prometheus-client==0.22.0
    # via -r requirements/requirements.in
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.31.0
    # via streamlit
pyarrow==20.0.0
    # via
    #   datasets
    #   streamlit
pydantic==2.11.5
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   pydantic-settings
    #   ragas
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via langchain-community
pydeck==0.9.1
    # via streamlit
pypdf==5.5.0
    # via -r requirements/requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via
    #   load-dotenv
    #   pydantic-settings
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   datasets
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
ragas==0.2.15
    # via -r requirements/requirements.in
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   datasets
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   streamlit
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.41
    # via
    #   langchain
    #   langchain-community
streamlit==1.45.1
    # via -r requirements/requirements.in
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
    #   streamlit
tiktoken==0.9.0
    # via
    #   langchain-openai
    #   ragas
toml==0.10.2
    # via streamlit
toolz==1.0.0
    # via altair
tornado==6.5.1
    # via streamlit
tqdm==4.67.1
    # via
    #   datasets
    #   huggingface-hub
    #   openai
typing-extensions==4.13.2
    # via
    #   anyio
    #   huggingface-hub
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   streamlit
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via requests
xxhash==3.5.0
    # via datasets
yarl==1.20.0
    # via aiohttp
zstandard==0.23.0
    # via langsmith
